Promote sizes to a larger data type to ensure that table calculations are correct

+++ programs/tiled_flat_resolution/main.cpp
@@ -225,8 +225,8 @@ int main(int argc, char **argv){
   std::string flip_style = argv[5];
 
   if(argv[2]==std::string("table")){
-    int dtype_size = GDALGetDataTypeSizeBytes(file_type);
-    int tile_size  = peekLayoutTileSize(argv[1]);
+    long dtype_size = GDALGetDataTypeSizeBytes(file_type);
+    long tile_size  = peekLayoutTileSize(argv[1]);
     for(int i=2;i<500;i++)
       std::cerr<<std::setw(2)<<i<<"  "<<((dtype_size+sizeof(flowdirs_t))*tile_size*i/1000000.0)<<" MB"<<std::endl;
     return -1;
